+++ axelor-production/src/main/java/com/axelor/apps/production/db/repo/ProductionBatchManagementRepository.java
@@ -15,15 +15,16 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.axelor.apps.production.report;
+package com.axelor.apps.production.db.repo;
 
-public interface IReport {
+import com.axelor.apps.production.db.ProductionBatch;
 
-  public static final String MANUF_ORDER = "ManufOrder.rptdesign";
-  public static final String OPERATION_ORDER = "OperationOrder.rptdesign";
-  public static final String PROD_PROCESS = "ProdProcess.rptdesign";
-  public static final String BILL_OF_MATERIAL = "BillOfMaterial.rptdesign";
-  public static final String RAW_MATERIAL_REQUIREMENT = "RawMaterialRequirement.rptdesign";
-  public static final String COST_SHEET = "CostSheet.rptdesign";
-  public static final String WORK_IN_PROGRESS_VALUATION = "WorkInProgressValuation.rptdesign";
+public class ProductionBatchManagementRepository extends ProductionBatchRepository {
+
+  @Override
+  public ProductionBatch copy(ProductionBatch entity, boolean deep) {
+    ProductionBatch copy = super.copy(entity, deep);
+    copy.setBatchList(null);
+    return copy;
+  }
 }