+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/StockMoveLineService.java
@@ -146,7 +146,8 @@ public interface StockMoveLineService {
       throws AxelorException;
 
   public void updateAveragePriceLocationLine(
-      StockLocation stockLocation, StockMoveLine stockMoveLine, int fromStatus, int toStatus);
+      StockLocation stockLocation, StockMoveLine stockMoveLine, int fromStatus, int toStatus)
+      throws AxelorException;
 
   /**
    * Check in the product if the stock move line needs to have a conformity selected.