+++ CHANGELOG.md
@@ -153,6 +153,7 @@ If not date is filled, added an option to choose or not to include it for sale o
 - Added mass payment on supplier invoices and generate account move on bank order validation.
 - Created CostSheet report.Added button to print Cost sheet report from Manufacturing order form.
 - Added management of subscription moves with MoveTemplate object
+- Batch : Add 'Production Batch : compute work in progress valuation' and generate PDF in production module.
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.
@@ -244,6 +245,7 @@ Only display line where quantity is greater than 0.
 - PURCHASE ORDER : Allow to edit delivery informations on validated purchase order.
 - Stock Move : Specify views for origin field. 
 - STOCK MOVE : Set avilableStatus in stockMoveLineList.
+- Add 'stock unit' field on stock location line and add unit conversion.
 
 ## Bug Fixes
 - Check duplicate : Apply manual filter before duplicate check and allow to open form view.